SFDM-2553  Создать метод POST api/v1/business-capability/public/{id}

+++ src/main/java/ru/beeline/capability/repository/BusinessCapabilityRepository.java
@@ -47,6 +47,6 @@ public interface BusinessCapabilityRepository extends JpaRepository<BusinessCapa
 
     Page<BusinessCapability> findByIsDomainTrue(Pageable pageable);
 
-    Optional<BusinessCapability> findByIdAndDeletedDateIsNullAndIsDomainTrue(Integer id);
+    Optional<BusinessCapability> findByIdAndDeletedDateIsNullAndIsDomainTrue(Long id);
 
 }
